refactor : moving scope scheme to util layer
