+++ highlight/src/main/java/com/neo/highlight/util/scheme/OnClickScheme.java
@@ -6,16 +6,12 @@ import android.view.View;
 
 import androidx.annotation.ColorInt;
 import androidx.annotation.NonNull;
-import androidx.annotation.Nullable;
 
 import com.neo.highlight.core.Scheme;
 import com.neo.highlight.core.ScopeScheme;
 import com.neo.highlight.util.scheme.base.BaseScheme;
 import com.neo.highlight.util.scheme.contract.LinkSchemeContract;
 
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.List;
 import java.util.regex.Pattern;
 
 /**
