+++ highlight/src/main/java/com/neo/highlight/util/scheme/StyleScheme.java
@@ -4,15 +4,9 @@ import android.graphics.Typeface;
 import android.text.style.StyleSpan;
 
 import androidx.annotation.NonNull;
-import androidx.annotation.Nullable;
 
-import com.neo.highlight.core.Scheme;
-import com.neo.highlight.core.ScopeScheme;
 import com.neo.highlight.util.scheme.base.BaseScheme;
 
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.List;
 import java.util.regex.Pattern;
 
 /**
